refactor: Update rest user dashboard routes
