+++ frontend/js/src/explore/art-creator/ArtCreator.tsx
@@ -23,7 +23,7 @@ import Preview from "./components/Preview";
 import ToggleOption from "./components/ToggleOption";
 import { svgToBlob, toPng } from "./utils";
 import { ToastMsg } from "../../notifications/Notifications";
-import UserSearch from "../../playlists/UserSearch";
+import UserSearch from "../../user/playlists/components/UserSearch";
 
 export enum TemplateNameEnum {
   designerTop5 = "designer-top-5",
