+++ frontend/js/src/user/recommendations/components/RecommendationPlaylistSettings.tsx
@@ -8,10 +8,10 @@ import {
 } from "@fortawesome/free-solid-svg-icons";
 import { sanitize } from "dompurify";
 import NiceModal from "@ebay/nice-modal-react";
-import { getPlaylistExtension, getPlaylistId } from "../playlists/utils";
-import { preciseTimestamp } from "../utils/utils";
-import GlobalAppContext from "../utils/GlobalAppContext";
-import ListenPayloadModal from "../listens/ListenPayloadModal";
+import { getPlaylistExtension, getPlaylistId } from "../../../playlists/utils";
+import { preciseTimestamp } from "../../../utils/utils";
+import GlobalAppContext from "../../../utils/GlobalAppContext";
+import ListenPayloadModal from "../../../listens/ListenPayloadModal";
 
 export type RecommendationPlaylistSettingsProps = {
   playlist: JSPFPlaylist;
