+++ frontend/js/src/user/taste/UserTaste.tsx
@@ -6,14 +6,14 @@ import * as Sentry from "@sentry/react";
 
 import { Integrations } from "@sentry/tracing";
 import NiceModal from "@ebay/nice-modal-react";
-import GlobalAppContext from "../utils/GlobalAppContext";
-import withAlertNotifications from "../notifications/AlertNotificationsHOC";
+import GlobalAppContext from "../../utils/GlobalAppContext";
+import withAlertNotifications from "../../notifications/AlertNotificationsHOC";
 
-import BrainzPlayer from "../brainzplayer/BrainzPlayer";
-import ErrorBoundary from "../utils/ErrorBoundary";
-import { getListenablePin, getPageProps } from "../utils/utils";
-import UserFeedback from "./UserFeedback";
-import UserPins from "../pins/UserPins";
+import BrainzPlayer from "../../brainzplayer/BrainzPlayer";
+import ErrorBoundary from "../../utils/ErrorBoundary";
+import { getListenablePin, getPageProps } from "../../utils/utils";
+import UserFeedback from "../UserFeedback";
+import UserPins from "../../pins/UserPins";
 
 export type UserTasteProps = {
   feedback?: Array<FeedbackResponseWithTrackMetadata>;
