+++ frontend/js/src/user/year-in-music/2021/YearInMusic2021.tsx
@@ -26,7 +26,7 @@ import BrainzPlayer from "../../../brainzplayer/BrainzPlayer";
 import withAlertNotifications from "../../../notifications/AlertNotificationsHOC";
 
 import { getPageProps } from "../../../utils/utils";
-import { getEntityLink } from "../../../stats/utils";
+import { getEntityLink } from "../../stats/utils";
 import ComponentToImage from "./components/ComponentToImage";
 
 import ListenCard from "../../../listens/ListenCard";
