+++ frontend/js/src/user/year-in-music/2022/YearInMusic2022.tsx
@@ -37,14 +37,14 @@ import {
   generateAlbumArtThumbnailLink,
   getPageProps,
 } from "../../../utils/utils";
-import { getEntityLink } from "../../../stats/utils";
+import { getEntityLink } from "../../stats/utils";
 import MagicShareButton from "./components/MagicShareButton";
 
 import ListenCard from "../../../listens/ListenCard";
 import UserListModalEntry from "../../../follow/UserListModalEntry";
 import { JSPFTrackToListen } from "../../../playlists/utils";
 import { COLOR_LB_ORANGE } from "../../../utils/constants";
-import CustomChoropleth from "../../../stats/Choropleth";
+import CustomChoropleth from "../../stats/components/Choropleth";
 import { ToastMsg } from "../../../notifications/Notifications";
 
 export type YearInMusicProps = {
