+++ frontend/js/src/user/year-in-music/2023/YearInMusic2023.tsx
@@ -41,13 +41,13 @@ import {
   getPageProps,
   getStatsArtistLink,
 } from "../../../utils/utils";
-import { getEntityLink } from "../../../stats/utils";
+import { getEntityLink } from "../../stats/utils";
 import ImageShareButtons from "./components/ImageShareButtons";
 
 import ListenCard from "../../../listens/ListenCard";
 import UserListModalEntry from "../../../follow/UserListModalEntry";
 import { JSPFTrackToListen } from "../../../playlists/utils";
-import CustomChoropleth from "../../../stats/Choropleth";
+import CustomChoropleth from "../../stats/components/Choropleth";
 import { ToastMsg } from "../../../notifications/Notifications";
 import FollowButton from "../../../follow/FollowButton";
