+++ frontend/js/tests/missing-mb-data/MissingMBData.test.tsx
@@ -5,7 +5,7 @@ import { act } from "react-dom/test-utils";
 import * as missingDataProps from "../__mocks__/missingMBDataProps.json";
 import { youtube, spotify, user } from "../__mocks__/missingMBDataProps.json";
 
-import MissingMBDataPage from "../../src/missing-mb-data/MissingMBData";
+import MissingMBDataPage from "../../src/user/missing-data/MissingMBData";
 import GlobalAppContext, {
   GlobalAppContextT,
 } from "../../src/utils/GlobalAppContext";
